Changed uint to size_t in idcat for MSVC++ compliance.
Former-commit-id: caa9041d3042b17c2b59503ec3aa2f8a8e78ecf6

+++ pwiz_tools/commandline/idcat.cpp
@@ -205,7 +205,7 @@ Config parseCommandLine(int argc, const char* argv[])
     {
         // For now, add 'em all.
         Serializer_Text::IdField last = Serializer_Text::Last;
-        for (uint i=1; i<(uint)last; i++)
+        for (size_t i=1; i<(size_t)last; i++)
         {
             config.fields.push_back(Serializer_Text::IdFieldNames[i]);
         }
